Use heuristics to define `time_t` instead of importing from system headers (#8)
Since `time_t` definition is not guaranteed in the C specification, CFFI needs to be told what it is.

However, each platform defines it in different headers in in different ways. This makes it very hard to import the definition from a system header (in a platform independent way) without also importing a bunch other unneeded definitions, which break `pycparser` (used by CFFI). 

The currrent solution to this problem was to just delete the unneeded header noise from the preprocessed system headers using regex expressios and header/line allowlists. However, this is very flimsy and easily breaks on different platforms.

This PR changes the behaviour so that instead of trying to parse out the `time_t` definition from system headers, it is computed during the FFI generation (see `ffi_generator.py`).  For Python 3.12+ this is relatively easy, since `ctypes.c_time_t` can be used to determine the correct C type. For Python 3.11 and below however, no such definition exists, thus the definition generation is based on heuristics instead.